+++ src/JSystem/JKernel/JKRArchivePub.cpp
@@ -38,7 +38,7 @@ JKRArchive* JKRArchive::mount(const char* path, EMountMode mountMode, JKRHeap* h
 
 JKRArchive* JKRArchive::mount(void* ptr, JKRHeap* heap,
                               EMountDirection mountDirection) {
-    JKRArchive* archive = check_mount_already((s32)ptr, heap);
+    JKRArchive* archive = check_mount_already((intptr_t)ptr, heap);
     if (archive) {
         return archive;
     }
